fit: Commit 1.4. First startup, fix bugs in boot/host/server/package and rework Secret-Code system to int

+++ src/main/java/tvgirl/elmodev/e1m0Admin/api/repo/gui/SecretCodeRepositoryAPI.java
@@ -5,11 +5,12 @@ import java.util.UUID;
 public interface SecretCodeRepositoryAPI {
 
     /* STAFF | 🧑‍🔬 */
-    void staffSetSecretCode(UUID adminID, UUID staffID, byte code);
-    void staffChangeSecretCode(UUID adminID, UUID staffID, byte code);
+    void staffSetSecretCode(UUID adminID, UUID staffID, int code);
+
+    void staffChangeSecretCode(UUID adminID, UUID staffID, int code);
 
     /* SYSTEM | 💾 */
-    void systemSetSecretCode(UUID adminID, byte code);
+    void systemSetSecretCode(UUID adminID, int code);
     byte getSecretCode(UUID adminID);
 
 }
